+++ src/ploomber_core/telemetry/telemetry.py
@@ -426,6 +426,17 @@ def validate_entries(event_id, uid, action, client_time, total_runtime):
     return event_id, uid, action, client_time, elapsed_time
 
 
+def is_first_arg_self(sig):
+    """
+    Check the func is defined inside a class
+
+    1. If the self as its first argument, it's a method
+    2. Otherwise, it's a function
+    """
+    params = list(sig.parameters)
+    return len(params) > 0 and params[0] == "self"
+
+
 class TelemetryGroup:
     def __init__(self, telemetry, group) -> None:
         self._telemetry = telemetry
@@ -647,26 +658,65 @@ class Telemetry:
         3
 
 
+        Unit testing (check the ``_telemetry`` attribute):
 
+        >>> from ploomber_core.telemetry import Telemetry
+        >>> telemetry = Telemetry("APIKEY", "packagename", "0.1")
+        >>> @telemetry.log_call()
+        ... def add(x, y):
+        ...     return x + y
+        >>> add._telemetry["action"]
+        'packagename-add'
+        >>> add._telemetry["payload"]
+        False
+        >>> add._telemetry["log_args"]
+        False
+        >>> add._telemetry["ignore_args"]
+        set()
+        >>> add._telemetry["group"]
         """
+
         if ignore_args is None:
             ignore_args = set()
         else:
             ignore_args = set(ignore_args)
 
         def _log_call(func):
+            # we'll use this on each call, so compute it once
+            func._signature = signature(func)
+
+            is_method = is_first_arg_self(func._signature)
+            # determine action name
+            action_ = self.package_name
+
+            if group:
+                action_ = f"{action_}-{group}"
+
+            name = action or getattr(func, "__name__", "funcion-without-name")
+            action_ = (f"{action_}-{name}").replace("_", "-")
+
+            func._telemetry_started = None
+            func._telemetry_success = None
+            func._telemetry_error = None
+
+            # store data for unit testing decorated functions
+            func._telemetry = dict(
+                action=action_,
+                payload=payload,
+                log_args=log_args,
+                ignore_args=ignore_args,
+                group=group,
+            )
+
             @wraps(func)
             def wrapper(*args, **kwargs):
-                action_ = self.package_name
-
-                if group:
-                    action_ = f"{action_}-{group}"
-
-                name = action or getattr(func, "__name__", "funcion-without-name")
-                action_ = (f"{action_}-{name}").replace("_", "-")
+                # reset attributes before calling
+                func._telemetry_started = None
+                func._telemetry_success = None
+                func._telemetry_error = None
 
                 if log_args:
-                    args_parsed = _get_args(func, args, kwargs, ignore_args)
+                    args_parsed = _get_args(func._signature, args, kwargs, ignore_args)
                 else:
                     args_parsed = None
 
@@ -677,12 +727,20 @@ class Telemetry:
                 if log_args:
                     metadata_started["args"] = args_parsed
 
-                self.log_api(action=f"{action_}-started", metadata=metadata_started)
+                started = dict(action=f"{action_}-started", metadata=metadata_started)
+                func._telemetry_started = started
+                self.log_api(**started)
+
                 start = datetime.datetime.now()
 
                 try:
                     if payload:
-                        result = func(_payload, *args, **kwargs)
+                        if is_method:
+                            injected_args = list(args)
+                            injected_args.insert(1, _payload)
+                            result = func(*injected_args, **kwargs)
+                        else:
+                            result = func(_payload, *args, **kwargs)
                     else:
                         result = func(*args, **kwargs)
                 except Exception as e:
@@ -697,11 +755,13 @@ class Telemetry:
                     if log_args:
                         metadata_error["args"] = args_parsed
 
-                    self.log_api(
+                    error = dict(
                         action=f"{action_}-error",
                         total_runtime=str(datetime.datetime.now() - start),
                         metadata=metadata_error,
                     )
+                    func._telemetry_error = error
+                    self.log_api(**error)
                     raise
                 else:
                     metadata_success = {"argv": get_sanitized_argv(), **_payload}
@@ -709,11 +769,13 @@ class Telemetry:
                     if log_args:
                         metadata_success["args"] = args_parsed
 
-                    self.log_api(
+                    success = dict(
                         action=f"{action_}-success",
                         total_runtime=str(datetime.datetime.now() - start),
                         metadata=metadata_success,
                     )
+                    func._telemetry_success = success
+                    self.log_api(**success)
 
                 return result
 
@@ -725,8 +787,8 @@ class Telemetry:
         return TelemetryGroup(self, group)
 
 
-def _get_args(func, fn_args, fn_kwargs, ignore_args):
-    mapping = _map_parameters_in_fn_call(fn_args, fn_kwargs, func)
+def _get_args(sig, fn_args, fn_kwargs, ignore_args):
+    mapping = _map_parameters_in_fn_call_from_signature(fn_args, fn_kwargs, sig)
 
     values_to_log = {}
 
@@ -767,7 +829,7 @@ def _process_value(value):
 
 
 # taken from sklearn-evaluation/util.py
-def _map_parameters_in_fn_call(args, kwargs, func):
+def _map_parameters_in_fn_call_from_signature(args, kwargs, sig):
     """
     Based on function signature, parse args to to convert them to key-value
     pairs and merge them with kwargs
@@ -775,7 +837,6 @@ def _map_parameters_in_fn_call(args, kwargs, func):
     is still passed.
     Missing parameters are filled with their default values
     """
-    sig = signature(func)
     # Get missing parameters in kwargs to look for them in args
     args_spec = list(sig.parameters)
     params_all = set(args_spec)